Fix home route build artifact path for Vercel
